+++ src/GlobalPool/GResourcePool.java
@@ -1,0 +1,5 @@
+package GlobalPool;
+
+public class GResourcePool {
+    
+}
