--- src/GlobalPool/GResourcesPool.java
@@ -1,5 +1,0 @@
-package GlobalPool;
-
-public class GResourcesPool {
-    
-}
