+++ src/PrefixState/Prefix.java
@@ -53,10 +53,10 @@ public enum Prefix {
     // Declare enum constants here.
     private String prefix;
     Prefix(String prefix) { this.prefix = prefix; }
-    Prefix(Object prefix) { this.prefix = (String) prefix; }
 
     public String GetPrefix() { return this.prefix; }
     public String GetPrefixCode() { return this.prefix; }
+    public String GetPrefixCodeTerm() { return this.prefix.replace("-", ""); }
 
     // ----------------------------------------------------------
     // Magic here: Declare the enum constant as a static final field using the prefix.
@@ -89,7 +89,6 @@ public enum Prefix {
         }
         return p;
     }
-
     public static boolean FindEnum(String inputString) { return Prefix.GetEnum(inputString) != null; }
 
 }
