+++ src/Room/LResourcePool.java
@@ -1,0 +1,5 @@
+package Room;
+
+public class LResourcePool {
+    
+}
