--- src/Room/LResourcesPool.java
@@ -1,5 +1,0 @@
-package Room;
-
-public class LResourcesPool {
-    
-}
