+++ src/Room/RoomUnit.java
@@ -2,7 +2,6 @@ package Room;
 
 import BaseClass.IntermediateObject;
 import java.util.Hashtable;
-import java.util.function.Function;
 
 import org.apache.commons.lang3.StringUtils;
 import PrefixState.Prefix;
@@ -70,8 +69,6 @@ import PrefixState.Prefix;
 **/
 
 public class RoomUnit extends IntermediateObject {
-    private static final Prefix prefix = PrefixState.Prefix.RoomUnit;
-
     public RoomUnit(String ID, String name, String description) {
         super(ID, name, description);
     }
@@ -79,212 +76,27 @@ public class RoomUnit extends IntermediateObject {
     // ---------------------------------------------------------------------------------------------------------------------
     // Getter and Setter
     public String GetRoomCodeID() { return this.GetID(); }                  // Alias function of GetID()
-    public static Prefix GetPrefix() { return RoomUnit.prefix; }
-    public static String GetPrefixCode() { return RoomUnit.GetPrefix().GetPrefixCode(); }
-    public static String GetPrefixCodeTerm() { return RoomUnit.GetPrefix().GetPrefixCode().replace("-", ""); }
+    public static Prefix GetPrefix() { return RoomUnitUtils.GetPrefix(); }
+    public static String GetPrefixCode() { return RoomUnitUtils.GetPrefix().GetPrefixCode(); }
+    public static String GetPrefixCodeTerm() { return RoomUnitUtils.GetPrefix().GetPrefixCodeTerm(); }
 
     // ---------------------------------------------------------------------------------------------------------------------
     // Extracting the information of the room-code
     public static String[] DecomposeRoomCodeID(String RoomCodeID) { return RoomCodeID.split("-"); }
 
-    public static String GetRoomBlock(String RoomCodeID) {return RoomUnit.DecomposeRoomCodeID(RoomCodeID)[1]; }
-    public String GetRoomBlock() { return RoomUnit.GetRoomBlock(this.GetID()); }
+    public String GetRoomBlock() { return RoomUnitUtils.GetRoomBlock(this.GetID()); }
 
-    public static String GetRoomType(String RoomCodeID) {return RoomUnit.DecomposeRoomCodeID(RoomCodeID)[2]; }
     public static RoomUnitEnum GetRoomTypeEnum(String RoomCodeID) { 
-        return RoomUnitEnum.GetEnum(RoomUnit.GetRoomType(RoomCodeID)); 
+        return RoomUnitEnum.GetEnum(RoomUnitUtils.GetRoomType(RoomCodeID)); 
     }
-    public String GetRoomType() { return RoomUnit.GetRoomType(this.GetID()); }
-    public RoomUnitEnum GetRoomTypeEnum() { return RoomUnit.GetRoomTypeEnum(this.GetID()); }
+    public RoomUnitEnum GetRoomTypeEnum() { return RoomUnitUtils.GetRoomTypeEnum(this.GetID()); }
 
-    public static String GetRoomFloor(String RoomCodeID) {return RoomUnit.DecomposeRoomCodeID(RoomCodeID)[3]; }
-    public String GetRoomFloor() { return RoomUnit.GetRoomFloor(this.GetID()); }
-
-    public static String GetRoomNumber(String RoomCodeID) {return RoomUnit.DecomposeRoomCodeID(RoomCodeID)[4]; }
-    public String GetRoomNumber() { return RoomUnit.GetRoomNumber(this.GetID()); }
+    public String GetRoomFloor() { return RoomUnitUtils.GetRoomFloor(this.GetID()); }
+    public String GetRoomNumber() { return RoomUnitUtils.GetRoomNumber(this.GetID()); }
 
     // -----------------------------------------------------------
     // Verifying the room-code
-    private static boolean VerifyRoomBlock(String RoomBlock) {
-        // The <room-block> can have at most two characters either numeric or alphabetical.
-        // If the <room-block> is a number, then it must be a two digit-like character.
-        // If the <room-block> is a character, then it must have one or two characters.
-        if (RoomBlock.length() > 2 || RoomBlock.length() == 0) { return false; }
-        
-        if (RoomBlock.length() == 1) {
-            char c = RoomBlock.charAt(0);
-            return Character.isLetter(c) && Character.isUpperCase(c);
-        }
-        
-        char c1 = RoomBlock.charAt(0);
-        char c2 = RoomBlock.charAt(1);
-        if (Character.isDigit(c1)) { return Character.isDigit(c2); }
-        if (Character.isLetter(c1) && Character.isUpperCase(c1)) { 
-            return Character.isLetter(c2) && Character.isUpperCase(c2); 
-        }
-        return false;
-    }
-
-    private static boolean VerifyRoomType(String RoomType) {
-        // The room-type is strongly dependent on the functionalities of the hospital room. 
-        // So the length of the room-type is not limited. However, we can predict and set 
-        // the constraint of the symbol's length. In this code, we choose the constraint to 
-        // be at most five (5) characters. As long it is a common letter (upper-case or 
-        // lower-case is acceptable), it can passed this test.
-        if (RoomType.length() > 5) { return false; }
-        for (int i = 0; i < RoomType.length(); i++) {
-            if (!Character.isLetter(RoomType.charAt(i))) { return false; }
-        }
-        return true;
-    }
-
-    private static boolean VerifyRoomFloor(String RoomFloor) {
-        if (RoomFloor.length() != 2) { return false; }
-        if (!Character.isDigit(RoomFloor.charAt(0))) { return false; }
-        if (!Character.isDigit(RoomFloor.charAt(1))) { return false; }
-        return true;
-    }
-
-    private static boolean VerifyRoomNumber(String RoomNumber) {
-        if (RoomNumber.length() != 3) { return false; }
-        if (!Character.isDigit(RoomNumber.charAt(0))) { return false; }
-        if (!Character.isDigit(RoomNumber.charAt(1))) { return false; }
-        if (!Character.isDigit(RoomNumber.charAt(2))) { return false; }
-        return true;
-    }
-
-    private static boolean VerifyRoomCodeID(String RoomCodeID) {
-        // This function is to verify the room-code ID as a single identity.
-        // As mentioned in the documentation above, we decompose the room-code ID into
-        // multiple parts using delimitor "-" and then verify each part.
-        String[] DecomposedRoomCodeID = RoomUnit.DecomposeRoomCodeID(RoomCodeID);
-        if (DecomposedRoomCodeID.length != 5) { return false; }
-        if (!RoomUnit.GetPrefixCodeTerm().equals(DecomposedRoomCodeID[0])) { return false; }
-        if (!RoomUnit.VerifyRoomBlock(RoomUnit.GetRoomBlock(RoomCodeID))) { return false; }
-        if (!RoomUnit.VerifyRoomType(RoomUnit.GetRoomType(RoomCodeID))) { return false; }
-        if (!RoomUnit.VerifyRoomFloor(RoomUnit.GetRoomFloor(RoomCodeID))) { return false; }
-        if (!RoomUnit.VerifyRoomNumber(RoomUnit.GetRoomNumber(RoomCodeID))) { return false; }
-        return true;
-    }
-    
-    public boolean VerifyRoomCodeID() { return RoomUnit.VerifyRoomCodeID(this.GetID()); }
-
-    // -----------------------------------------------------------
-    // Constructing or Generating the Room-Code
-    // Note that the most complex part is to guaranteed that each part of the room-code ID
-    // must be canonicalized or standardized. For example, the room-block must be in upper-case.
-    // Note that this class is the low-level so guarantee the uniqueness amongst the room-code
-    // ID is not necessary, as this job is for the RoomManager.java.
-    // Note that if each part of the room-code ID is not satisfied the easiest condition, then 
-    // it is invalid, which returns the null or raise IllegalArgumentException.
-    // Since the RoomBlock, RoomFloor, and RoomNumber can be an integer, we have to program another 
-    // method with different signature. With this integer method, you can cast back to String, 
-    // call the respective method; but it is not necessary and this call is extremely faster.
-
-    private static String CastRoomBlock(String RoomBlock) {
-        // This method will cast your RoomBlock to have two characters, except if the input is
-        // the letter-like character with one character only.
-        if (RoomBlock == null) { return null; }
-        if (RoomBlock.length() > 2 || RoomBlock.length() == 0) { return null; }
-        if (RoomBlock.length() == 1) {
-            char c = RoomBlock.charAt(0);
-            
-            // String.format("%02d", Integer.parseInt(Character.toString(c)));
-            if (Character.isDigit(c)) { return "0" + String.valueOf(c); } 
-            if (Character.isLetter(c)) { return RoomBlock.toUpperCase(); }
-            return null;
-        }
-
-        if (RoomBlock.length() == 2) {
-            char c1 = RoomBlock.charAt(0);
-            char c2 = RoomBlock.charAt(1);
-
-            if (Character.isDigit(c1) && Character.isDigit(c2)) { return RoomBlock; }
-            if (Character.isLetter(c1) && Character.isLetter(c2)) { return RoomBlock.toLowerCase(); }
-            return null;
-        }
-        return null;
-
-    }
-
-    private static String CastRoomBlock(int RoomBlock) {
-        if (RoomBlock <= 0 || RoomBlock >= 99) { return null; }
-        // You can cast to string, call RoomUnit.CastRoomBlock(String) and then cast back to int.
-        // However, it is not necessary and this call is extremely faster.
-        return String.format("%02d", RoomBlock); 
-    }
-
-    private static String CastRoomType(String RoomType) {
-        // This method will cast your RoomType to have at most five characters.
-        if (RoomType == null) { return null; }
-        if (RoomType.length() > 5) { return null; }
-        for (int i = 0; i < RoomType.length(); i++) {
-            if (!Character.isLetter(RoomType.charAt(i))) { return null; }
-        }
-        return RoomType.toUpperCase();
-    }
-
-    private static String CastRoomFloor(String RoomFloor) {
-        // This method will cast your RoomFloor to have two digit-like characters.
-        if (RoomFloor == null) { return null; }
-        if (RoomFloor.length() == 0 || RoomFloor.length() > 2) { return null; }
-        for (int i = 0; i < RoomFloor.length(); i++) {
-            if (!Character.isDigit(RoomFloor.charAt(i))) { return null; }
-        }
-        return String.format("%02d", Integer.parseInt(RoomFloor));
-    }
-
-    private static String CastRoomFloor(int RoomFloor) {
-        if (RoomFloor <= 0 || RoomFloor >= 99) { return null; }
-        return String.format("%02d", RoomFloor);
-    }
-
-    private static String CastRoomNumber(String RoomNumber) {
-        // This method will cast your RoomNumber to have three digit-like characters.
-        if (RoomNumber == null) { return null; }
-        if (RoomNumber.length() == 0 || RoomNumber.length() > 3) { return null; }
-        for (int i = 0; i < RoomNumber.length(); i++) {
-            if (!Character.isDigit(RoomNumber.charAt(i))) { return null; }
-        }
-        return String.format("%03d", Integer.parseInt(RoomNumber));
-    }
-
-    private static String CastRoomNumber(int RoomNumber) {
-        if (RoomNumber <= 0 || RoomNumber >= 999) { return null; }
-        return String.format("%03d", RoomNumber);
-    }
-
-    public static String ConstructRoomCodeID(String RoomBlock, String RoomType, String RoomFloor, String RoomNumber) {
-        // This function is to construct the room-code ID as a single identity.
-        // As mentioned in the documentation above, we construct the room-code ID by
-        // concatenating the parts of the room-code ID.
-        String RoomBlockStr, RoomTypeStr, RoomFloorStr, RoomNumberStr;
-         
-        if (StringUtils.isNumeric((CharSequence) RoomBlock)) { 
-            RoomBlockStr = RoomUnit.CastRoomBlock(Integer.parseInt(RoomBlock)); 
-        } else { RoomBlockStr = RoomUnit.CastRoomBlock(RoomBlock); }
-
-        RoomTypeStr = RoomUnit.CastRoomType(RoomType);
-
-        if (StringUtils.isNumeric((CharSequence) RoomFloor)) { 
-            RoomFloorStr = RoomUnit.CastRoomFloor(Integer.parseInt(RoomFloor)); 
-        } else { RoomFloorStr = RoomUnit.CastRoomFloor(RoomFloor); }
-
-        if (StringUtils.isNumeric((CharSequence) RoomNumber)) { 
-            RoomNumberStr = RoomUnit.CastRoomNumber(Integer.parseInt(RoomNumber)); 
-        } else { RoomNumberStr = RoomUnit.CastRoomNumber(RoomNumber); }
-
-        if (RoomBlockStr == null || RoomTypeStr == null || RoomFloorStr == null || RoomNumberStr == null) { 
-            String FullMessage = "";
-            if (RoomBlockStr == null) { FullMessage += "Invalid RoomBlock: " + RoomBlock.toString() + "; "; }
-            if (RoomTypeStr == null) { FullMessage += "Invalid RoomType: " + RoomType + "; "; }
-            if (RoomFloorStr == null) { FullMessage += "Invalid RoomFloor: " + RoomFloor.toString() + "; "; }
-            if (RoomNumberStr == null) { FullMessage += "Invalid RoomNumber: " + RoomNumber.toString() + "; "; }
-            throw new IllegalArgumentException(FullMessage);
-        }
-        String[] iterable = {RoomBlockStr, RoomTypeStr, RoomFloorStr, RoomNumberStr};
-        return RoomUnit.GetPrefixCodeTerm() + StringUtils.join(iterable, "-");
-    }
+    public boolean VerifyRoomCodeID() { return RoomUnitUtils.VerifyRoomCodeID(this.GetID()); }
 
 
     // ---------------------------------------------------------------------------------------------------------------------
